requirements.txt
panda
matplotlib
plotly
geopandas
dash